+++ src/main/java/aero/cubox/cmmn/controller/CommonController.java
@@ -86,38 +86,6 @@ public class CommonController {
 		}
 	}
 
-	@SuppressWarnings("serial")
-	@RequestMapping(value="/main.do")
-	public String main(ModelMap model, @RequestParam Map<String, Object> commandMap, RedirectAttributes redirectAttributes, HttpSession session) throws Exception {
-		//자동새로고침때문에 추가
-		String reloadYn = StringUtil.isNullToString(commandMap.get("reloadYn")).matches("Y") ? StringUtil.isNullToString(commandMap.get("reloadYn")) : "N";
-		String intervalSecond = StringUtil.isNullToString(commandMap.get("intervalSecond")).matches("(^[0-9]+$)") ? StringUtil.isNullToString(commandMap.get("intervalSecond")) : "5";
-
-		model.addAttribute("reloadYn", reloadYn);
-		model.addAttribute("intervalSecond", intervalSecond);
-
-		String ssAuthorId = ((LoginVO)session.getAttribute("loginVO")).getRole_id();
-
-
-
-
-
-		//사용자가 해당 메뉴(근태관리)에 접근 권한이 있는지 조회
-//		List<MenuDetailVO> result =  menuService.getUserMenuList(new HashMap<String, Object>() {
-//			{
-//				put("role_id", ssAuthorId);
-//				put("menu_cl_code", "00009");//근태관리 메뉴 코드
-//			}
-//		});
-		
-		//model.addAttribute("isMenu", (result != null && result.size() > 0) ? true : false );
-
-
-
-		
-
-		return "cubox/common/main";
-	}
