+++ src/main/java/aero/cubox/cmmn/service/MainService.java
@@ -3,6 +3,7 @@ package aero.cubox.cmmn.service;
 import aero.cubox.core.vo.*;
 
 import java.util.List;
+import java.util.Map;
 
 
 public interface MainService {
@@ -13,4 +14,5 @@ public interface MainService {
 
 	int getDayEntEmpCount() throws Exception;
 
+    List<Map> getMainAlarmHistList() throws Exception;
 }
